- Seperated training code from the Diffusion class (renamed as GaussianDiffusion in a new file denoising_diffusion_outlier.py) - Also moved the UNet model instantiation to the GaussianDiffusion class, which is inline with lucidrains' Diffusion code (denoising_diffusion_pytorch/). - Also add option to load from a checkpoint (see --ckpt) and do sampling from it.
